+++ gargantua/models/__init__.py
@@ -1,8 +1,1 @@
-from .base import BaseBlogModel
-from .articles import ArticlesModel, ArticleBufferModel
-
-
-__all__ = [
-    'BaseBlogModel',
-    'ArticlesModel', 'ArticleBufferModel',
-]
+from .articles import BaseBlogModel, ArticlesModel, ArticleBufferModel